refactor(pdp): move PDP-related code do pdp directory closer to PDP views
