+++ apps/storefront/src/pdp/components/product-details-skeleton.tsx
