+++ apps/storefront/src/pdp/components/related-products-skeleton.tsx
