+++ apps/storefront/src/pdp/components/variant-selector-utils.ts
